pyproject.toml
[project]
name = "colonoscopy-triage-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
